requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pandas==2.3.3
pydantic==2.12.5
importlib-metadata==8.7.1
python-dotenv==1.2.1
google-genai
requests==2.32.5
mcp==1.25.0
fastmcp==2.14.2
pyngrok==7.5.0
python-multipart==0.0.21
